make converter components modular
